feat: separate functions into separte files to make them modular
